2-add-settlement_spot-class Refactor test files
